implemented error handling for wrong values
