+++ geo/basic.py
@@ -2,56 +2,45 @@ import turtle
 
 blk = (0,0,0)
 
-def rect(height : float, length : float, fill, source):
-    t = turtle.Turtle()
+def rect(height: int, length: int, fill: bool, source):
+    rt = turtle.Turtle()
+    rt.hideturtle()
+    rt.penup()
+    rt.goto(0, 0)
+    rt.setheading(0)
+    rt.pendown()
+    rt.speed(0) 
+    if fill and source:
+        rt.color(source)
+        rt.begin_fill()
+    else:
+        rt.color(source if source else blk)
+    for _ in range(2):
+        rt.forward(length)
+        rt.left(90)
+        rt.forward(height)
+        rt.left(90)
+    if fill and source:
+        rt.end_fill()
+
+
+
+def carr(cote : int, fill, source):
+    ct = turtle.Turtle()
+    ct.speed(0)
     if source == None:
-        t.color(blk)
+        ct.color(blk)
     elif fill == True and bool(source) == True:
         fcolor = source
-        t.color(fcolor)
-        t.begin_fill()
-
-    t.forward(length)
-    t.left(90)
-    t.forward(height)
-    t.left(90)
-    t.forward(length)
-    t.left(90)
-    t.forward(height)
-    if fill:
-        t.end_fill()
-
-
-def carr(cote : float, fill, source):
-    t = turtle.Turtle()
-    if source == None:
-        t.color(blk)
-    elif fill == True and bool(source) == True:
-        fcolor = source
-        t.color(fcolor)
-        t.begin_fill()
+        ct.color(fcolor)
+        ct.begin_fill()
 
     for x in range(0, 5):
-        t.forward(cote)
-        t.left(90)
+        ct.forward(cote)
+        ct.left(90)
     if fill == True:
-        t.end_fill()
+        ct.end_fill()
 
-
-def triang(cote : float, fill, source):
-    t = turtle.Turtle()
-    if source == None:
-        t.color(blk)
-    elif fill == True and bool(source) == True:
-        fcolor = source
-        t.color(fcolor)
-        t.begin_fill()
-
-    for x in range(1, 4):
-        t.forward(cote)
-        t.left(360 / 3)
-    if fill == True:
-        t.end_fill()
 
 
 def circle(rad, fill, source):
