+++ tests.md
@@ -1,0 +1,29 @@
+autoDrawGUI tests:
+    1.UI/ UX: yes
+        icons : yes
+        buttons: yes
+        sliders: yes
+        graying outs: yes
+    2.Shapes:
+        Rectangle: yes
+            outline:
+                color:yes
+                size: yes
+            fill: yes
+            measurements: yes
+        Square:
+            outline:yes
+                color:yes
+                size:yes
+            fill:yes
+            measurments:yes
+        triangle: (equilateral ONLY): yes
+            outline:yes
+                color:yes
+                size:yes
+            fill:yes
+            measurments:yes
+    3.Error Handling: yes
+        try/except: yes
+        msgboxes: yes
+        ZeroDivisionError: yes
